consider text and cell styles in shared strings
