+++ lib/creek.rb
@@ -1,12 +1,11 @@
 require "creek/version"
 require 'creek/book'
-require 'creek/html_extractor'
+require 'creek/cell_value_extractor'
 require 'creek/styles/constants'
 require 'creek/styles/style_types'
 require 'creek/styles/converter'
 require 'creek/styles'
 require 'creek/cell_styles'
-require 'creek/extractor'
 require 'creek/sheet'
 require 'creek/shared_strings'
