+++ lib/creek/shared_strings.rb
@@ -3,10 +3,13 @@ require 'nokogiri'
 
 module Creek
   class Creek::SharedStrings
+    include Creek::CellValueExtractor
+
     attr_reader :files, :options, :dictionary
 
     def initialize(files, options = {})
-      @files, @options = files, options
+      @files = files
+      @options = options
       @dictionary = parse_shared_strings
     end
 
@@ -17,7 +20,9 @@ module Creek
         doc = files.file.open path
         xml = Nokogiri::XML::Document.parse doc
 
-        @dictionary ||= xml.css('si').map { |si| Creek::Extractor.new(si, options).extract }
+        @dictionary ||= xml.css('si').map do |node|
+          options[:with_html] ? node : text_from(node)
+        end
       end
     end
   end
